+++ test/harness/cases/button_manners.lua
@@ -8,7 +8,7 @@ return function(H)
     "VerdantGraphWindowWelcomeOkBtn",
     "VerdantBarWindowPrevBtn", "VerdantBarWindowNextBtn", "VerdantBarWindowModeBtn",
     "VerdantBarWindowSettingsBtn", "VerdantBarWindowGraphBtn", "VerdantBarWindowCloseBtn",
-    "VerdantLibraryOpenBtn", "VerdantLibraryLockBtn", "VerdantLibraryDeleteBtn", "VerdantLibraryCloseBtn",
+    "VerdantLibraryOpenBtn", "VerdantLibraryLockBtn", "VerdantLibraryDeleteBtn", "VerdantLibraryCloseBtn", "VerdantLibraryLabelBtn",
     "VerdantSettingsPanelPSaveBtn", "VerdantSettingsPanelPDeleteBtn", "VerdantSettingsPanelAutoRecBtn",
     "VerdantSettingsPanelAutosaveBtn", "VerdantSettingsPanelLightBtn", "VerdantSettingsPanelShieldDirBtn",
     "VerdantSettingsPanelGdmBtn", "VerdantSettingsPanelUnknownBtn", "VerdantSettingsPanelLogoBtn",